expose interface

+++ contracts/libraries/helper/Math.sol
@@ -78,7 +78,8 @@ library Math {
             return result >= roundedDownResult ? roundedDownResult : result;
         }
     }
-    function min(uint x, uint y) internal pure returns (uint z) {
+
+    function min(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = x < y ? x : y;
     }
 }
